+++ src/datatable/HISTORY.md
@@ -4,6 +4,9 @@ DataTable Change History
 @VERSION@
 ------
 
+* Added 'contentUpdate' after the DataTable has been updated when triggered
+  from a `dataChange` event. [Pull Request #1072][Issue #1052]
+
 * Fix issue where recursive nesting of objects was cloned infinitely
   [Pull Request #1008][Ticket #915]